\name{rjson}
\docType{package}
\alias{rjson}
\alias{rjson-package}
\title{JSON for R}

\description{ JSON (JavaScript Object Notation) is a lightweight data-interchange format. 
This package converts JSON objects into R objects and vice-versa. See json.org for an
overview of JSON. Unicode is unfortunately not supported at the moment.
}

\keyword{package}
