\name{fromJSON}
\alias{fromJSON}
\title{Convert a R object into JSON notation}

\description{ Convert an R object into JSON notation }

\usage{fromJSON( json_str )}

\arguments{
\item{json_str}{JSON string notation}
}

\value{R object that corresponds to the JSON object}

\seealso{
\code{\link{fromJSON}}
}

\examples{
#TODO add example here
}
%TODO find better keyword
\keyword{math}
