\name{toJSON}
\alias{toJSON}
\title{Convert a R object into JSON notation}

\description{ Convert an R object into JSON notation }

\usage{toJSON( x )}

\arguments{
\item{x}{A vector or list}
}

\value{String containing the JSON notation of the R object}

\seealso{
\code{\link{toJSON}}
}

\examples{
#TODO add example here
}
%TODO find better keyword
\keyword{math}
